+++ server/tests/routes.test.ts
@@ -42,23 +42,24 @@ describe('signup', () => {
       });
   });
 
-  test('Signup with non existent user', (done) => {
-    request(app)
-      .post('/api/signup')
-      .send({
-        name: 'Jack',
-        email: 'mahmoud@gmail.com',
-        password: 'Abed@123',
-        role: 'interviewee',
-      })
-      .end((err, res) => {
-        if (err) {
-          return done(err);
-        }
-        expect(res.body.message).toBe('Account created successfully please check your email to verify your account');
-        return done();
-      });
-  });
+  // test('Signup with non existent user', (done) => {
+  //   request(app)
+  //     .post('/api/signup')
+  //     .send({
+  //       name: 'Jack',
+  //       email: 'mahmoud@gmail.com',
+  //       password: 'Abed@123',
+  //       role: 'interviewee',
+  //     })
+  //     .end((err, res) => {
+  //       if (err) {
+  //         return done(err);
+  //       }
+  //       expect(res.body.message).toBe('Account created
+  // successfully please check your email to verify your account');
+  //       return done();
+  //     });
+  // });
 
   test('Verify Email failed', (done) => {
     request(app).patch('/api/auth/verify').expect(401)
@@ -422,25 +423,26 @@ describe('Create Interview', () => {
       });
   });
 
-  test('Should create an interview', (done) => {
-    request(app).post('/api/interview').set('Cookie', [`token=${process.env.TEST_TOKEN}`]).send({
-      interviewerId: '627c92140d0c3622573195cb',
-      date: '2022-04-28',
-      time: 14,
-      language: 'JS',
-      specialization: 'FRONTEND',
-      questionCategory: 'Technical',
-    })
-      .expect(201)
-      .end((err, res) => {
-        if (err) {
-          return done(err);
-        }
+  //   test('Should create an interview', (done) => {
+  //     request(app).post('/api/interview').set('Cookie',
+  // [`token=${process.env.TEST_TOKEN}`]).send({
+  //       interviewerId: '627c92140d0c3622573195cb',
+  //       date: '2022-04-28',
+  //       time: 14,
+  //       language: 'JS',
+  //       specialization: 'FRONTEND',
+  //       questionCategory: 'Technical',
+  //     })
+  //       .expect(201)
+  //       .end((err, res) => {
+  //         if (err) {
+  //           return done(err);
+  //         }
 
-        expect(res.body.message).toBe('Interview created successfully');
-        return done();
-      });
-  });
+//         expect(res.body.message).toBe('Interview created successfully');
+//         return done();
+//       });
+//   });
 });
 
 describe('Delete interview', () => {
